+++ src/ReleaseNotes.md
@@ -1,8 +1,9 @@
 # Release Notes
-Minor release extending signing certificate support
+Streamlined option naming and improved usability
 
 ## Features
-- Extended the AzureCreateSigningCert tool to support creating signing certificates locally signed by a KeyVault based CA certificate.
+- Made ClientID and TenantID optional when using 'WorkloadIdentity'.
+- Unified Certificate name option to 'CertificateName' across all Azure tools for consistency.
 
 ## Bug Fixes
 - None